Remove large resolvers file and split into focused modules
Removed large backend/src/resolvers.js file and split code into separate files focusing on specific functionalities or entities. Implemented this change to make the codebase more manageable, clear, and scalable. This also facilitates separate maintenance and further feature development for different parts of the codebase.
