+++ backend/src/resolvers/IssueComment/index.js
@@ -1,0 +1,11 @@
+const resolvers = {
+  Query: {},
+  Mutation: {},
+  IssueComment: {
+    reporter: (parent, args, { db }) => {
+      return db.sequelize.models.User.findByPk(parent.reporterId);
+    },
+  },
+};
+
+export default resolvers;
