+++ backend/src/resolvers/Upload/index.js
@@ -1,0 +1,7 @@
+import GraphQLUpload from 'graphql-upload/GraphQLUpload.mjs';
+
+const resolvers = {
+  Upload: GraphQLUpload,
+};
+
+export default resolvers;
