--- frontend/.prettierrc.json
@@ -1,8 +1,0 @@
-{
-  "trailingComma": "es5",
-  "semi": true,
-  "tabWidth": 2,
-  "singleQuote": true,
-  "jsxSingleQuote": true,
-  "plugins": ["prettier-plugin-tailwindcss"]
-}
